// backend/src/generated/prisma/models/Dynasty.ts
export type DynastyModel = runtime.Types.Result.DefaultSelection<Prisma.$DynastyPayload>
// ...
export type $DynastyPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Dynasty"
  objects: {
    emperors: Prisma.$EmperorPayload<ExtArgs>[]
    tangFigures: Prisma.$TangFigurePayload<ExtArgs>[]
    songFigures: Prisma.$SongFigurePayload<ExtArgs>[]
    yuanFigures: Prisma.$YuanFigurePayload<ExtArgs>[]
    mingFigures: Prisma.$MingFigurePayload<ExtArgs>[]
    qingRulers: Prisma.$QingRulerPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    startYear: number
    endYear: number | null
    capital: string | null
    founder: string | null
    description: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["dynasty"]>
  composites: {}
}